+++ appGUI/preferences/cncjob/CNCJobPreferencesUI.py
@@ -4,7 +4,7 @@ from appGUI.preferences.cncjob.CNCJobAdvOptPrefGroupUI import CNCJobAdvOptPrefGr
 from appGUI.preferences.cncjob.CNCJobOptPrefGroupUI import CNCJobOptPrefGroupUI
 from appGUI.preferences.cncjob.CNCJobGenPrefGroupUI import CNCJobGenPrefGroupUI
 from appGUI.preferences.cncjob.CNCJobEditorPrefGroupUI import CNCJobEditorPrefGroupUI
-
+from appGUI.preferences.cncjob.CNCJobPPGroupUI import CNCJobPPGroupUI
 
 class CNCJobPreferencesUI(QtWidgets.QWidget):
 
@@ -26,9 +26,13 @@ class CNCJobPreferencesUI(QtWidgets.QWidget):
         self.cncjob_editor_group = CNCJobEditorPrefGroupUI(decimals=self.decimals, defaults=self.defaults)
         self.cncjob_editor_group.setMinimumWidth(260)
 
+        self.cncjob_pp_group = CNCJobPPGroupUI(decimals=self.decimals, defaults=self.defaults)
+        self.cncjob_pp_group.setMinimumWidth(260)
+
         vlay = QtWidgets.QVBoxLayout()
         vlay.addWidget(self.cncjob_opt_group)
         vlay.addWidget(self.cncjob_adv_opt_group)
+        vlay.addWidget(self.cncjob_pp_group)
 
         self.layout.addWidget(self.cncjob_gen_group)
         self.layout.addLayout(vlay)