+++ preprocessors/Default_no_M6.py
@@ -205,8 +205,20 @@ G00 Z{z_toolchange}
         return 'G01 Z0'
 
     def position_code(self, p):
+        # formula for skewing on x for example is:
+        # x_fin = x_init + y_init/slope where slope = p._bed_limit_y / p._bed_skew_x (a.k.a tangent)
+        if p._bed_skew_x == 0:
+            x_pos = p.x + p._bed_offset_x
+        else:
+            x_pos = (p.x + p._bed_offset_x) + ((p.y / p._bed_limit_y) * p._bed_skew_x)
+
+        if p._bed_skew_y == 0:
+            y_pos = p.y + p._bed_offset_y
+        else:
+            y_pos = (p.y + p._bed_offset_y) + ((p.x / p._bed_limit_x) * p._bed_skew_y)
+
         return ('X' + self.coordinate_format + ' Y' + self.coordinate_format) % \
-               (p.coords_decimals, p.x, p.coords_decimals, p.y)
+               (p.coords_decimals, x_pos, p.coords_decimals, y_pos)
 
     def rapid_code(self, p):
         return ('G00 ' + self.position_code(p)).format(**p)