+++ preprocessors/ISEL_ICP_CNC.py
@@ -143,7 +143,19 @@ class ISEL_ICP_CNC(PreProc):
         return 'MOVEABS Z0'
 
     def position_code(self, p):
-        return 'X' + str(int(p.x * 1000)) + ' Y' + str(int(p.y * 1000))
+        # formula for skewing on x for example is:
+        # x_fin = x_init + y_init/slope where slope = p._bed_limit_y / p._bed_skew_x (a.k.a tangent)
+        if p._bed_skew_x == 0:
+            x_pos = int((p.x * 1000) + p._bed_offset_x)
+        else:
+            x_pos = int((p.x * 1000) + p._bed_offset_x + (((p.y * 1000) / p._bed_limit_y) * p._bed_skew_x))
+
+        if p._bed_skew_y == 0:
+            y_pos = int((p.y * 1000) + p._bed_offset_y)
+        else:
+            y_pos = int((p.x * 1000) + p._bed_offset_x + (((p.x * 1000) / p._bed_limit_x) * p._bed_skew_y))
+
+        return 'X' + str(x_pos) + ' Y' + str(y_pos)
 
     def rapid_code(self, p):
         return ('FASTABS ' + self.position_code(p)).format(**p)